+++ official/wide_deep/census_dataset.py
@@ -18,12 +18,16 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import argparse
 import os
 import sys
 
+from absl import app as absl_app
+from absl import flags
 from six.moves import urllib
 import tensorflow as tf
+
+from official.utils.flags import core as flags_core
+
 
 DATA_URL = 'https://archive.ics.uci.edu/ml/machine-learning-databases/adult'
 TRAINING_FILE = 'adult.data'
@@ -31,11 +35,21 @@ TRAINING_URL = '%s/%s' % (DATA_URL, TRAINING_FILE)
 EVAL_FILE = 'adult.test'
 EVAL_URL = '%s/%s' % (DATA_URL, EVAL_FILE)
 
-parser = argparse.ArgumentParser()
 
-parser.add_argument(
-    '--data_dir', type=str, default='/tmp/census_data',
-    help='Directory to download census data')
+_CSV_COLUMNS = [
+  'age', 'workclass', 'fnlwgt', 'education', 'education_num',
+  'marital_status', 'occupation', 'relationship', 'race', 'gender',
+  'capital_gain', 'capital_loss', 'hours_per_week', 'native_country',
+  'income_bracket'
+]
+
+_CSV_COLUMN_DEFAULTS = [[0], [''], [0], [''], [0], [''], [''], [''], [''], [''],
+                        [0], [0], [0], [''], ['']]
+
+_NUM_EXAMPLES = {
+  'train': 32561,
+  'validation': 16281,
+}
 
 
 def _download_and_clean_file(filename, url):
@@ -55,17 +69,130 @@ def _download_and_clean_file(filename, url):
   tf.gfile.Remove(temp_file)
 
 
+def _download(data_dir):
+  if not tf.gfile.Exists(data_dir):
+    tf.gfile.MkDir(data_dir)
+
+  training_file_path = os.path.join(data_dir, TRAINING_FILE)
+  if not tf.gfile.Exists(training_file_path):
+    _download_and_clean_file(training_file_path, TRAINING_URL)
+
+  eval_file_path = os.path.join(data_dir, EVAL_FILE)
+  if not tf.gfile.Exists(eval_file_path):
+    _download_and_clean_file(eval_file_path, EVAL_URL)
+
+
+def build_model_columns():
+  """Builds a set of wide and deep feature columns."""
+  # Continuous columns
+  age = tf.feature_column.numeric_column('age')
+  education_num = tf.feature_column.numeric_column('education_num')
+  capital_gain = tf.feature_column.numeric_column('capital_gain')
+  capital_loss = tf.feature_column.numeric_column('capital_loss')
+  hours_per_week = tf.feature_column.numeric_column('hours_per_week')
+
+  education = tf.feature_column.categorical_column_with_vocabulary_list(
+      'education', [
+        'Bachelors', 'HS-grad', '11th', 'Masters', '9th', 'Some-college',
+        'Assoc-acdm', 'Assoc-voc', '7th-8th', 'Doctorate', 'Prof-school',
+        '5th-6th', '10th', '1st-4th', 'Preschool', '12th'])
+
+  marital_status = tf.feature_column.categorical_column_with_vocabulary_list(
+      'marital_status', [
+        'Married-civ-spouse', 'Divorced', 'Married-spouse-absent',
+        'Never-married', 'Separated', 'Married-AF-spouse', 'Widowed'])
+
+  relationship = tf.feature_column.categorical_column_with_vocabulary_list(
+      'relationship', [
+        'Husband', 'Not-in-family', 'Wife', 'Own-child', 'Unmarried',
+        'Other-relative'])
+
+  workclass = tf.feature_column.categorical_column_with_vocabulary_list(
+      'workclass', [
+        'Self-emp-not-inc', 'Private', 'State-gov', 'Federal-gov',
+        'Local-gov', '?', 'Self-emp-inc', 'Without-pay', 'Never-worked'])
+
+  # To show an example of hashing:
+  occupation = tf.feature_column.categorical_column_with_hash_bucket(
+      'occupation', hash_bucket_size=1000)
+
+  # Transformations.
+  age_buckets = tf.feature_column.bucketized_column(
+      age, boundaries=[18, 25, 30, 35, 40, 45, 50, 55, 60, 65])
+
+  # Wide columns and deep columns.
+  base_columns = [
+    education, marital_status, relationship, workclass, occupation,
+    age_buckets,
+  ]
+
+  crossed_columns = [
+    tf.feature_column.crossed_column(
+        ['education', 'occupation'], hash_bucket_size=1000),
+    tf.feature_column.crossed_column(
+        [age_buckets, 'education', 'occupation'], hash_bucket_size=1000),
+  ]
+
+  wide_columns = base_columns + crossed_columns
+
+  deep_columns = [
+    age,
+    education_num,
+    capital_gain,
+    capital_loss,
+    hours_per_week,
+    tf.feature_column.indicator_column(workclass),
+    tf.feature_column.indicator_column(education),
+    tf.feature_column.indicator_column(marital_status),
+    tf.feature_column.indicator_column(relationship),
+    # To show an example of embedding
+    tf.feature_column.embedding_column(occupation, dimension=8),
+  ]
+
+  return wide_columns, deep_columns
+
+
+def input_fn(data_file, num_epochs, shuffle, batch_size):
+  """Generate an input function for the Estimator."""
+  assert tf.gfile.Exists(data_file), (
+      '%s not found. Please make sure you have run census_dataset.py and '
+      'set the --data_dir argument to the correct path.' % data_file)
+
+  def parse_csv(value):
+    print('Parsing', data_file)
+    columns = tf.decode_csv(value, record_defaults=_CSV_COLUMN_DEFAULTS)
+    features = dict(zip(_CSV_COLUMNS, columns))
+    labels = features.pop('income_bracket')
+    return features, tf.equal(labels, '>50K')
+
+  # Extract lines from input files using the Dataset API.
+  dataset = tf.data.TextLineDataset(data_file)
+
+  if shuffle:
+    dataset = dataset.shuffle(buffer_size=_NUM_EXAMPLES['train'])
+
+  dataset = dataset.map(parse_csv, num_parallel_calls=5)
+
+  # We call repeat after shuffling, rather than before, to prevent separate
+  # epochs from blending together.
+  dataset = dataset.repeat(num_epochs)
+  dataset = dataset.batch(batch_size)
+  return dataset
+
+
+def define_data_download_flags():
+  """Add flags specifying data download arguments."""
+  flags.DEFINE_string(
+      name="data_dir", default="/tmp/census_data/",
+      help=flags_core.help_wrap(
+          "Directory to download and extract data."))
+
+
 def main(_):
-  if not tf.gfile.Exists(FLAGS.data_dir):
-    tf.gfile.MkDir(FLAGS.data_dir)
-
-  training_file_path = os.path.join(FLAGS.data_dir, TRAINING_FILE)
-  _download_and_clean_file(training_file_path, TRAINING_URL)
-
-  eval_file_path = os.path.join(FLAGS.data_dir, EVAL_FILE)
-  _download_and_clean_file(eval_file_path, EVAL_URL)
+  _download(flags.FLAGS.data_dir)
 
 
 if __name__ == '__main__':
-  FLAGS, unparsed = parser.parse_known_args()
-  tf.app.run(argv=[sys.argv[0]] + unparsed)
+  tf.logging.set_verbosity(tf.logging.INFO)
+  define_data_download_flags()
+  absl_app.run(main)
